Added MessagePrinter and Counter
Added classes MessagePrinter and Counter used to print messages in
standard output (cout) atomically and to count elements atomically
respectively.
The class server was modified in order to use this class instead of the
Server::showMessage to print and to count threads with this Counter.
Besides, the 'ShowMessage' macro uses the MessagePrinter now.